+++ includes/class-cocart-install.php
@@ -5,7 +5,7 @@
  * @author  Sébastien Dumont
  * @package CoCart\Classes
  * @since   1.2.0
- * @version 3.1.0
+ * @version 3.7.2
  * @license GPL-2.0+
  */
 
@@ -163,7 +163,7 @@ if ( ! class_exists( 'CoCart_Install' ) ) {
 		 * @access  public
 		 * @static
 		 * @since   1.2.0
-		 * @version 3.1.0
+		 * @version 3.7.2
 		 */
 		public static function install() {
 			if ( ! is_blog_installed() ) {
@@ -176,6 +176,12 @@ if ( ! class_exists( 'CoCart_Install' ) ) {
 
 			// Check if we are not already running this routine.
 			if ( 'yes' === get_transient( 'cocart_installing' ) ) {
+				return;
+			}
+
+			// If WooCommerce is NOT active, don't install CoCart.
+			if ( ! defined( 'WC_VERSION' ) ) {
+				CoCart_Admin_Notices::add_notice( 'check_wc' );
 				return;
 			}